This is a file where we can put the palette of colors we re gonna use
The colors i put are just an example we can change them later. I wanted to put them in one file and then refer to them by a name so that when we want to change a certain color we dont go around changing thel on by one. We can decide on the color palette whenever this way :)